uelearner: add code to learn mealy machine
This learner uses the w-method for equivalence checking (which is slow).
An alternative is to use the random work equivalence check, which is
less precise but faster.

The learner dumps the global variable accesses for the `GLOBAL_STATE`
variable, but this might not be sufficient to represent state in a
practical system, as this might have heap allocations that represent
state.

Signed-off-by: Eduard Vlad <eduard.vlad@rwth-aachen.de>