uelearner: add helper scripts for visualization and inconsistency checking
cleanup_model.py: groups same inputs that have the same s/t states and
same output on the edge. This way looking at the graph is way more
legible.

find_inconsistency.py: Assuming that the learner logs I: (*), O: (*),
this allows to do a sanity check against race conditions or similar
bugs. This way we can rule out indeterminism, which is not allowed for
active automata learning (it conflicts the requirements).

Signed-off-by: Eduard Vlad <eduard.vlad@rwth-aachen.de>